+++ packages/extensions/modelerfour/test/modeler/modelerfour.test.ts
@@ -3,20 +3,9 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-import {
-  Parameter,
-  SchemaResponse,
-  ConstantSchema,
-  SealedChoiceSchema,
-  DictionarySchema,
-  Operation,
-} from "@autorest/codemodel";
-import { addOperation, addSchema, createTestSpec, InitialTestSpec, response, responses } from "../utils";
+import { Parameter, SchemaResponse, ConstantSchema, SealedChoiceSchema } from "@autorest/codemodel";
+import { addOperation, addSchema, createTestSpec, findByName, InitialTestSpec, response, responses } from "../utils";
 import { runModeler } from "./modelerfour-utils";
-
-export function findByName<T>(name: string, items: Array<T> | undefined): T | undefined {
-  return (items && items.find((i) => (<any>i).language.default.name === name)) || undefined;
-}
 
 function assertSchema(
   schemaName: string,
@@ -867,82 +856,6 @@ describe("Modeler", () => {
     expect(nonApiVersionQueryParam!.language.default.serializedName).toEqual("api-version");
     expect(nonApiVersionQueryParam!.implementation).toEqual("Method");
     expect(nonApiVersionQueryParam!.origin).toEqual(undefined);
-  });
-
-  describe("x-ms-header-collection-prefix headers", () => {
-    let spec: any;
-    let hasHeaderWithExtension: Operation;
-
-    beforeEach(async () => {
-      spec = createTestSpec();
-
-      addOperation(spec, "/headerWithExtension", {
-        post: {
-          operationId: "hasHeaderWithExtension",
-          description: "Has x-ms-header-collection-prefix on header",
-          parameters: [
-            {
-              "name": "x-ms-req-meta",
-              "x-ms-client-name": "RequestHeaderWithExtension",
-              "in": "header",
-              "schema": {
-                type: "string",
-              },
-              "x-ms-parameter-location": "method",
-              "x-ms-header-collection-prefix": "x-ms-req-meta",
-            },
-          ],
-          responses: responses(
-            response(
-              200,
-              "application/json",
-              {
-                type: "string",
-              },
-              "Response with a header extension.",
-              {
-                headers: {
-                  "x-named-header": {
-                    "x-ms-client-name": "HeaderWithExtension",
-                    "x-ms-header-collection-prefix": "x-ms-res-meta",
-                    "schema": {
-                      type: "string",
-                    },
-                  },
-                },
-              },
-            ),
-          ),
-        },
-      });
-      const codeModel = await runModeler(spec);
-      hasHeaderWithExtension = findByName("hasHeaderWithExtension", codeModel.operationGroups[0].operations)!;
-      expect(hasHeaderWithExtension).not.toBeNull();
-    });
-
-    it("propagates extensions to response header definitions", async () => {
-      const headerWithExtension = hasHeaderWithExtension.responses?.[0].protocol.http!.headers[0];
-      expect(headerWithExtension.language.default.name).toEqual("HeaderWithExtension");
-      expect(headerWithExtension.extensions["x-ms-header-collection-prefix"]).toEqual("x-ms-res-meta");
-    });
-
-    it("changes the type of the response header to be Dictionary<originalType>", async () => {
-      const headerWithExtension = hasHeaderWithExtension.responses?.[0].protocol.http!.headers[0];
-      expect(headerWithExtension.schema.type).toEqual("dictionary");
-      expect(headerWithExtension.schema.elementType.type).toEqual("string");
-    });
-
-    it("propagates extensions to request header definitions", async () => {
-      const headerWithExtension = hasHeaderWithExtension.parameters?.[1];
-      expect(headerWithExtension?.language.default.name).toEqual("RequestHeaderWithExtension");
-      expect(headerWithExtension?.extensions?.["x-ms-header-collection-prefix"]).toEqual("x-ms-req-meta");
-    });
-
-    it("changes the type of the request header to be Dictionary<originalType>", async () => {
-      const headerWithExtension = hasHeaderWithExtension.parameters?.[1];
-      expect(headerWithExtension?.schema.type).toEqual("dictionary");
-      expect((headerWithExtension?.schema as DictionarySchema).elementType.type).toEqual("string");
-    });
   });
 
   it("allows text/plain responses when schema type is 'string'", async () => {
