+++ packages/extensions/modelerfour/test/utils/codemodel-utils.ts
@@ -1,0 +1,3 @@
+export function findByName<T>(name: string, items: Array<T> | undefined): T | undefined {
+  return (items && items.find((i) => (<any>i).language.default.name === name)) || undefined;
+}
