+++ packages/extensions/modelerfour/test/utils/index.ts
@@ -1,3 +1,4 @@
+export * from "./codemodel-utils";
 export * from "./test-session";
 export * from "./specs";
 export * from "./schema-utils";
